+++ src/devices/sharedFunctions/index.ts
@@ -2,4 +2,5 @@ export * from './battery';
 export * from './handleSensor';
 export * from './humiditySensor';
 export * from './lampUtils';
+export * from './shutterUtils';
 export * from './temperatureSensor';