+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Centralise some shutter functions
+
 ## 3.2.9 (2025-07-02)
 
 * (theimo1221) Some fixes for Victron Energy Manager